Refactors EL_parcel_locations and EL_parcel_value_assessment dags to be DRYer.
